requirements.txt
pandas
numpy
scikit-learn
tensorflow
pillow
pyheif
ultralytics